pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sas-py-demo"
version = "0.1.0"
description = "CLI demo for sas-py: streams microphone and webcam to the SD Attention Server and routes detected speech through OpenAI Realtime."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "sas-py>=0.1.0",
]
